--- utbot-python-executor/src/test/kotlin/org/utbot/python/executor/TestDeepSerialization.kt
@@ -1,12 +1,0 @@
-package org.utbot.python.executor
-
-import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.api.Test
-
-internal class DeepSerializationTest {
-
-    @Test
-    fun testUtBotExecutor() {
-        assertTrue(true)
-    }
-}
