+++ utbot-python/src/main/kotlin/org/utbot/python/evaluation/serialization/ExecutionResultDeserializer.kt
@@ -1,4 +1,4 @@
-package org.utbot.python.evaluation.serialiation
+package org.utbot.python.evaluation.serialization
 
 import com.squareup.moshi.JsonEncodingException
 import com.squareup.moshi.Moshi
